save() becomes a delegate to the moved method
save(PasswordDatabase) should be moved to the PasswordDatabase class as it refers to more methods and fields of that class. encryptionService will be passed as a parameter in the new method and fields FILE_HEADER and DB_VERSION need to be public so that they can be accessed by the body of the moved method.